Ya se eneseñan las portadas de los albumes y la tabla ya hace refresh

+++ Music/Database/DataBase.cs
@@ -165,7 +165,7 @@ namespace Database
                 using(SQLiteCommand command = new SQLiteCommand(query, connection))
                 {
                     command.Parameters.AddWithValue("@id_performer", performer.IdPerformer);
-                    command.Parameters.AddWithValue("@id_type", performer.Type);
+                    command.Parameters.AddWithValue("@id_type", (int)performer.Type);
                     command.Parameters.AddWithValue("@name", performer.Name);
                     int rowsAffected = command.ExecuteNonQuery();
